Refactor refactor, add additional objects
