+++ include/socketserver.h
@@ -28,6 +28,7 @@
 // History:     Oct-26-2018     Davepl      Created
 //---------------------------------------------------------------------------
 #pragma once
+
 #include <unistd.h>
 #include <stdio.h>
 #include <sys/socket.h>
